# psd-render-api/openapi-current.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSD 渲染服务平台 API",
    "description": "## 概述\nPSD 模板渲染服务平台对外提供 PSD 模板管理、资产上传、渲染任务提交与查询的 HTTP API。\n\n## 鉴权\n所有 `/v1/*` 接口需在 `Authorization` 头携带 API Key（格式：`Bearer sk_live_xxx`）。\nAPI Key 通过 Admin 控制台创建，每个 Key 可配置：\n- 限流（rateLimitPerMin）：每分钟最大请求数\n- 日配额（quotaPerDay）：每天最大请求数\n- 作用域（scopes）：限定可调用的接口\n- IP 白名单（ipWhitelist）\n- 默认 Webhook URL / Secret\n\n## 幂等性\n`POST /v1/render-jobs` 支持通过 `Idempotency-Key` 头实现幂等：\n同租户同 key 重复提交将返回相同任务，不重复创建。\n\n## Webhook\n任务终态（succeeded / failed / cancelled）时自动投递 Webhook 到提交时声明的 `webhookUrl`。\n投递包含 HMAC-SHA256 签名头 `X-Render-Signature`，可校验来源。\n失败自动重试，指数退避：1m / 5m / 15m / 60m / 240m / 960m。\n\n## 错误响应\n所有错误响应统一为 `{ \"error\": \"<CODE>\", \"message\": \"<描述>\", \"details\"?: [...] }`。",
    "version": "0.3.0",
    "contact": {
      "name": "PSD 渲染服务平台"
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key（格式：sk_live_xxx）",
        "bearerFormat": "sk_live_xxx"
      },
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin Session Token（登录后获得）",
        "bearerFormat": "session_token"
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Worker 访问令牌（注册后获得）",
        "bearerFormat": "wkr_xxx"
      },
      "storageSignedToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "存储签名令牌（格式：Bearer <token>，由 generateUploadUrl/generateDownloadUrl 签发，仅 local 模式可用。也支持 ?token= 查询参数，但不推荐）"
      },
      "workerRegisterSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Worker 注册凭据（格式：Bearer <WORKER_REGISTER_SECRET> 或一次性配对码/token，也可通过 X-Worker-Register-Secret 头传递）"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "details": {
            "type": "array",
            "description": "可选，详细错误信息"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      }
    }
  },
  "paths": {
...
    "/admin/api/api-keys": {
      "get": {
        "summary": "API Key 列表",
        "tags": [
          "admin-api-keys"
        ],
        "description": "查询 API Key 列表，可按 tenantId / active 过滤。响应不含 keyHash（仅元数据）。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "按租户过滤"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "active",
            "required": false,
            "description": "按启用状态过滤"
          }
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKeys": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "keyPrefix": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "tenantId": {
                            "type": "string"
                          },
                          "active": {
                            "type": "boolean"
                          },
                          "priority": {
                            "type": "integer"
                          },
                          "rateLimitPerMin": {
                            "type": "integer",
                            "nullable": true
                          },
                          "quotaPerDay": {
                            "type": "integer",
                            "nullable": true
                          },
                          "quotaUsedDay": {
                            "type": "integer"
                          },
                          "quotaResetAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "scopes": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "webhookUrlDefault": {
                            "type": "string",
                            "nullable": true
                          },
                          "webhookSecret": {
                            "type": "string",
                            "nullable": true
                          },
                          "ipWhitelist": {
                            "type": "string",
                            "nullable": true
                          },
                          "disabledAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "lastUsedAt": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建 API Key",
        "tags": [
          "admin-api-keys"
        ],
        "description": "创建新的 API Key。\n\n**重要**：响应中的 `plaintextKey` 仅此一次返回，前端必须立即提示用户保存，后续无法再次获取。\n\n可配置项：\n- 限流（rateLimitPerMin）\n- 日配额（quotaPerDay）\n- 作用域（scopes）\n- IP 白名单（ipWhitelist）\n- 默认 Webhook URL / Secret",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "description": "API Key 名称"
                  },
                  "tenantId": {
                    "type": "string",
                    "maxLength": 64,
                    "description": "租户 ID（可选）"
                  },
                  "priority": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "description": "优先级（数字越小越高，默认 5）"
                  },
                  "rateLimitPerMin": {
                    "type": "integer",
                    "nullable": true,
                    "description": "每分钟最大请求数（null 表示不限）"
                  },
                  "quotaPerDay": {
                    "type": "integer",
                    "nullable": true,
                    "description": "每天最大请求数（null 表示不限）"
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "作用域列表"
                  },
                  "webhookUrlDefault": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true,
                    "description": "默认 Webhook URL"
                  },
                  "webhookSecret": {
                    "type": "string",
                    "maxLength": 256,
                    "nullable": true,
                    "description": "Webhook 签名密钥"
                  },
                  "ipWhitelist": {
                    "type": "string",
                    "maxLength": 1024,
                    "nullable": true,
                    "description": "IP 白名单（逗号分隔 CIDR/IP）"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "responses": {
          "201": {
            "description": "创建成功，plaintextKey 仅此一次返回",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "创建成功，plaintextKey 仅此一次返回",
                  "properties": {
                    "apiKey": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "plaintextKey": {
                      "type": "string",
                      "description": "明文 API Key（仅此一次返回，格式 sk_live_xxx）"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api/api-keys/{id}/rotate": {
      "post": {
        "summary": "轮换 API Key",
        "tags": [
          "admin-api-keys"
        ],
        "description": "生成新的明文 key 并替换 keyHash，旧 key 立即失效。响应中的 `plaintextKey` 仅此一次返回。operator 及以上可调用。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "API Key ID"
          }
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "plaintextKey": {
                      "type": "string",
                      "description": "新明文 API Key（仅此一次返回）"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/admin/api/api-keys/{id}/disable": {
      "post": {
        "summary": "禁用 API Key",
        "tags": [
          "admin-api-keys"
        ],
        "description": "将 API Key 设为 active=false，立即拒绝所有后续请求。operator 及以上可调用。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/admin/api/api-keys/{id}/enable": {
      "post": {
        "summary": "启用 API Key",
        "tags": [
          "admin-api-keys"
        ],
        "description": "将 API Key 设为 active=true，恢复接受请求。operator 及以上可调用。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/admin/api/api-keys/{id}/reset-quota": {
      "post": {
        "summary": "重置 API Key 日配额",
        "tags": [
          "admin-api-keys"
        ],
        "description": "将 API Key 当日已用配额清零，立即恢复可用额度。operator 及以上可调用。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "API Key ID"
          }
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}